fix filters representative to work on every basic object
